requirements.txt
-e git://github.com/dissemin/django-allauth.git@used_by_dissemin#egg=django-allauth
celery[redis]
pyoai
PyPDF2
bibtexparser>=1.1.0
redis<3
cryptography
django==1.11.20
django-autocomplete-light
django-bootstrap-pagination
django-bulk-update
django-bootstrap3-datepicker
django-cache-machine
django-capture-tag
django-cas-ng
django-countries
django-crispy-forms
django-geojson
django-haystack
django-js-reverse
django-jsonview
django-leaflet
django-memoize
django-redis-cache
django-solo
django-ratelimit
django-widget-tweaks
django-debug-toolbar
djangorestframework
elasticsearch
html5lib==0.999999999
httplib2
jsonfield
lxml
mock
multiprocessing_generator
name_tools
ndg-httpsclient
pillow
psycopg2
pyasn1
pyopenssl>=0.14
requests[security]
requests_mock
titlecase
unicode_tex
unidecode
urllib3
wand
raven
python-dateutil # for haystack, as explained here:
#http://django-haystack.readthedocs.io/en/v2.4.1/management_commands.html
